refactor: reorganise io package into io/read and io/write concern packages
All reading machinery (readers, builders, strategies, match patterns,
ClassPatternMapping, FixedFormatLineProcessor, MultiMatchStrategy) moves
to io/read. FixedFormatWriter moves to io/write. TypedReadResult stays
in io/ as the shared result type; io/row/ is unchanged.

This eliminates the MultiMatchStrategy workaround from the previous
refactor: with ClassPatternMapping now in io/read, MultiMatchStrategy
can co-locate with the rest of the reading infrastructure.

TypedReadResult constructor made public to allow construction from io/read.
616 tests green.
