+++ fixedformat4j/src/main/java/com/ancientprogramming/fixedformat4j/io/read/PackedRecordReaderBuilder.java
@@ -13,14 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ancientprogramming.fixedformat4j.io;
+package com.ancientprogramming.fixedformat4j.io.read;
 
 import com.ancientprogramming.fixedformat4j.format.FixedFormatManager;
 import com.ancientprogramming.fixedformat4j.format.impl.FixedFormatManagerImpl;
-import com.ancientprogramming.fixedformat4j.io.pattern.FixedFormatMatchPattern;
-import com.ancientprogramming.fixedformat4j.io.strategy.ParseErrorStrategy;
-import com.ancientprogramming.fixedformat4j.io.strategy.PartialChunkStrategy;
-import com.ancientprogramming.fixedformat4j.io.strategy.UnmatchStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
