+++ fixedformat4j/src/main/java/com/ancientprogramming/fixedformat4j/io/read/RegexFixedFormatMatchPattern.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ancientprogramming.fixedformat4j.io.pattern;
+package com.ancientprogramming.fixedformat4j.io.read;
 
 import com.ancientprogramming.fixedformat4j.exception.FixedFormatException;
