+++ fixedformat4j/src/main/java/com/ancientprogramming/fixedformat4j/io/write/FixedFormatWriter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ancientprogramming.fixedformat4j.io;
+package com.ancientprogramming.fixedformat4j.io.write;
 
 import com.ancientprogramming.fixedformat4j.exception.FixedFormatIOException;
 import com.ancientprogramming.fixedformat4j.format.FixedFormatManager;
@@ -56,7 +56,7 @@ import java.util.Objects;
  *
  * @author Jacob von Eyben - <a href="https://eybenconsult.com">https://eybenconsult.com</a>
  * @since 1.9.0
- * @see FixedFormatReader#readAsRows(File)
+ * @see com.ancientprogramming.fixedformat4j.io.read.FixedFormatReader#readAsRows(java.io.File)
  */
 public class FixedFormatWriter {
