+++ fixedformat4j/src/test/java/com/ancientprogramming/fixedformat4j/io/TestFixedFormatReaderBuilder.java
@@ -3,16 +3,19 @@ package com.ancientprogramming.fixedformat4j.io;
 import com.ancientprogramming.fixedformat4j.annotation.Field;
 import com.ancientprogramming.fixedformat4j.annotation.Record;
 import com.ancientprogramming.fixedformat4j.format.impl.FixedFormatManagerImpl;
-import com.ancientprogramming.fixedformat4j.io.pattern.FixedFormatMatchPattern;
-import com.ancientprogramming.fixedformat4j.io.pattern.RegexFixedFormatMatchPattern;
-import com.ancientprogramming.fixedformat4j.io.strategy.ParseErrorStrategy;
-import com.ancientprogramming.fixedformat4j.io.strategy.UnmatchStrategy;
+import com.ancientprogramming.fixedformat4j.io.read.FixedFormatMatchPattern;
+import com.ancientprogramming.fixedformat4j.io.read.RegexFixedFormatMatchPattern;
+import com.ancientprogramming.fixedformat4j.io.read.ParseErrorStrategy;
+import com.ancientprogramming.fixedformat4j.io.read.UnmatchStrategy;
 import org.junit.jupiter.api.Test;
 
 import java.io.StringReader;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
+import com.ancientprogramming.fixedformat4j.io.read.FixedFormatReader;
+import com.ancientprogramming.fixedformat4j.io.read.MultiMatchStrategy;
+import com.ancientprogramming.fixedformat4j.io.read.FixedFormatReaderBuilder;
 
 class TestFixedFormatReaderBuilder {
