+++ fixedformat4j/src/test/java/com/ancientprogramming/fixedformat4j/io/TestFixedFormatRoundTrip.java
@@ -1,8 +1,8 @@
 package com.ancientprogramming.fixedformat4j.io;
 
 import com.ancientprogramming.fixedformat4j.format.impl.FixedFormatManagerImpl;
-import com.ancientprogramming.fixedformat4j.io.pattern.FixedFormatMatchPattern;
-import com.ancientprogramming.fixedformat4j.io.pattern.RegexFixedFormatMatchPattern;
+import com.ancientprogramming.fixedformat4j.io.read.FixedFormatMatchPattern;
+import com.ancientprogramming.fixedformat4j.io.read.RegexFixedFormatMatchPattern;
 import com.ancientprogramming.fixedformat4j.io.row.ParsedRow;
 import com.ancientprogramming.fixedformat4j.io.row.Row;
 import com.ancientprogramming.fixedformat4j.io.row.UnmatchedRow;
@@ -17,6 +17,8 @@ import java.nio.file.Path;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
+import com.ancientprogramming.fixedformat4j.io.read.FixedFormatRowReader;
+import com.ancientprogramming.fixedformat4j.io.write.FixedFormatWriter;
 
 class TestFixedFormatRoundTrip {
